backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.1.0
python-dateutil>=2.9.0.post0
minio>=7.2.0

httpx>=0.27.0
beautifulsoup4>=4.12.0
pandas>=2.2.0
openpyxl>=3.1.2
